requirements/development.txt
# This file collects all required third-party applications that are needed
# to develop this project. Later you can install all these apps in a row
# using pip. Example::
#
#     pip install -r development.txt

Fabric>=1.5
django-discoverage>=0.5
mock
factory_boy